requirements.txt
asgiref==3.3.1
Django==3.1.4
django-ckeditor==6.0.0
django-js-asset==1.2.2
mysqlclient==2.0.2
Pillow==8.0.1
python-dotenv==0.15.0
pytz==2020.4
sqlparse==0.4.1
